moving the current select2_controller.js into folder v1 and adding component tests
