+++ app/components/viral/form/prefixed/select2_component.html.erb
@@ -12,7 +12,7 @@
     </div>
 
     <div class="grow">
-      <div data-action="keydown->viral--select2#keydown" class="select2">
+      <div data-action="keydown->viral--v1--select2#keydown" class="select2">
         <%= tag.input(
           type: "text",
           id: id,
@@ -26,21 +26,21 @@
             text-slate-800 sm:text-sm @min-md:rounded-l-none! dark:border-slate-600 dark:bg-slate-800
             dark:text-slate-50 dark:placeholder-slate-400
           ",
-          data: { "viral--select2-target": "input", action: "input->viral--select2#input" },
+          data: { "viral--v1--select2-target": "input", action: "input->viral--v1--select2#input" },
           aria: aria
         ) %>
 
-        <%= form.hidden_field name, data: { "viral--select2-target": "hidden" } %>
+        <%= form.hidden_field name, data: { "viral--v1--select2-target": "hidden" } %>
 
         <div
-          data-viral--select2-target="dropdown"
+          data-viral--v1--select2-target="dropdown"
           class="
             absolute z-10 hidden rounded-lg border border-slate-300 bg-white shadow-lg dark:border-slate-700
             dark:bg-slate-800
           "
         >
           <ul
-            data-viral--select2-target="scroller"
+            data-viral--v1--select2-target="scroller"
             role="listbox"
             class="
               max-h-[300px] w-full divide-y divide-slate-200 overflow-y-auto py-2 wrap-anywhere
@@ -60,7 +60,7 @@
                   pointer-events-none hidden border border-blue-300 bg-blue-50 p-4 text-blue-800
                   dark:border-blue-800 dark:bg-slate-800 dark:text-blue-400
                 "
-                data-viral--select2-target="empty"
+                data-viral--v1--select2-target="empty"
               >
                 <%= empty_state %>
               </li>
