+++ app/javascript/controllers/viral/v1/select2_controller.js
