+++ app/views/groups/_edit_advanced_transfer_form.html.erb
@@ -3,7 +3,7 @@ data:
 if Flipper.enabled?(:v2_select2)
     { turbo_confirm: "", controller: "viral--v2--select2" }
   else
-    { turbo_confirm: "", controller: "viral--select2" }
+    { turbo_confirm: "", controller: "viral--v1--select2" }
   end) do |form| %>
   <%= render partial: "shared/form/required_field_legend" %>
 
@@ -68,7 +68,7 @@ if Flipper.enabled?(:v2_select2)
                   turbo_content: ".confirm-transfer",
                   confirm_value: confirm_value,
                 } else {
-                  "viral--select2-target": "submitButton",
+                  "viral--v1--select2-target": "submitButton",
                   turbo_content: ".confirm-transfer",
                   confirm_value: confirm_value,
                 }
