+++ app/views/samples/clones/_dialog.html.erb
@@ -4,7 +4,7 @@
 
   <%= turbo_frame_tag "clone_samples_dialog_content" do %>
     <div
-      data-controller="infinite-scroll viral--select2"
+      data-controller="infinite-scroll viral--v1--select2"
       data-infinite-scroll-selection-outlet="#samples-table"
       data-infinite-scroll-paged-field-name-value="sample_ids[]"
       data-infinite-scroll-singular-description-value="<%= t("samples.clones.dialog.description.singular") %>"
@@ -118,7 +118,7 @@
                           {
                             action: "viral--dialog#setClosable",
                             "viral--dialog-closable-param": "false",
-                            "viral--select2-target": "submitButton",
+                            "viral--v1--select2-target": "submitButton",
                           }
                           end %>
             </div>
