+++ test/components/viral/form/prefixed/select2_component_test.rb
@@ -11,15 +11,15 @@ module Viral
           assert_selector 'input[type="hidden"][name="user"]', visible: :hidden, count: 1
 
           find('input.select2-input[type="text"]').click
-          assert_selector 'div[data-viral--select2-target="dropdown"]', visible: :visible
-          assert_selector 'ul[data-viral--select2-target="scroller"] li', count: 50
+          assert_selector 'div[data-viral--v1--select2-target="dropdown"]', visible: :visible
+          assert_selector 'ul[data-viral--v1--select2-target="scroller"] li', count: 50
 
           find('li[data-label="User 1"]').click
           assert_selector 'input[type="hidden"][name="user"][value="1"]', visible: :hidden, count: 1
 
           find('input.select2-input[type="text"]').send_keys([:ctrl, 'a'], '22')
-          assert_selector 'div[data-viral--select2-target="dropdown"]', visible: :visible
-          find('ul[data-viral--select2-target="scroller"] > li').click
+          assert_selector 'div[data-viral--v1--select2-target="dropdown"]', visible: :visible
+          find('ul[data-viral--v1--select2-target="scroller"] > li').click
 
           assert_selector 'input[type="hidden"][name="user"][value="22"]', visible: :hidden, count: 1
         end
