+++ test/system/projects_test.rb
@@ -34,7 +34,7 @@ class ProjectsTest < ApplicationSystemTestCase
     click_link I18n.t('general.navbar.new_dropdown.project')
 
     within %(div[data-controller="slugify"][data-controller-connected="true"]) do
-      assert_selector %(input[data-viral--select2-target="input"]) do |input|
+      assert_selector %(input[data-viral--v1--select2-target="input"]) do |input|
         assert_equal @user.namespace.full_path, input['value']
       end
     end
@@ -46,7 +46,7 @@ class ProjectsTest < ApplicationSystemTestCase
     click_link I18n.t('groups.show.create_project_button')
 
     within %(div[data-controller="slugify"][data-controller-connected="true"]) do
-      assert_selector %(input[data-viral--select2-target="input"]) do |input|
+      assert_selector %(input[data-viral--v1--select2-target="input"]) do |input|
         assert_equal group1.path, input['value']
       end
     end
